refactor(storybook-addon): update package structure
