+++ tooling/storybook-addon/package.json
@@ -16,20 +16,15 @@
       "require": "./dist/index.js"
     },
     "./*": "./*",
-    "./preset": {
-      "types": "./dist/preset/index.d.ts",
-      "import": "./dist/preset/index.mjs",
-      "require": "./dist/preset/index.js"
+    "./manager": {
+      "types": "./dist/manager.d.ts",
+      "require": "./dist/manager.js",
+      "import": "./dist/manager.mjs"
     },
-    "./preset/decorators": {
-      "types": "./dist/preset/index.d.ts",
-      "import": "./dist/preset/decorators.mjs",
-      "require": "./dist/preset/decorators.js"
-    },
-    "./preset/register": {
-      "types": "./dist/preset/index.d.ts",
-      "import": "./dist/preset/register.mjs",
-      "require": "./dist/preset/register.js"
+    "./preview": {
+      "types": "./dist/preview.d.ts",
+      "require": "./dist/preview.js",
+      "import": "./dist/preview.mjs"
     }
   },
   "typesVersions": {
