+++ tooling/storybook-addon/src/index.ts
@@ -1,2 +1,2 @@
 export * from "./constants"
-export * from "./feature/arg-types"
+export * from "./arg-types"
