--- tooling/storybook-addon/src/preset/decorators.ts
@@ -1,3 +1,0 @@
-import { ChakraProviderDecorator } from "../feature/decorator/ChakraProviderDecorator"
-
-export const decorators: any[] = [ChakraProviderDecorator]
