+++ tooling/storybook-addon/src/preview.ts
@@ -1,0 +1,3 @@
+import { ChakraProviderDecorator } from "./ChakraProviderDecorator"
+
+export const decorators: any[] = [ChakraProviderDecorator]
